ShoppingCartUsingJs
In this Shopping Cart page three products have been added to the basket
and the names of these are shown together with their cost and the
quantity desired for purchase.
Beneath the product list, a sub-total of the product costs are shown,
VAT is then added (at 20%) and the total is shown beneath.
Finally, a ‘Buy Now’ button allows the user to go ahead and purchase the
items shown at the total price given.
When there and products in the basket and the ‘Buy Now’ button is
selected, you should take the product data together with the quantities,
costs and totals, and simulate how you would POST this data in JSON
format to an Ajax-supported web service URL.